Removed obsolete code for installing server_redhat packages with explicit names.
I suspect this resulted from a merge accident. This code was originally removed
in the 2012-08-06 commit for "[COOK-1532] - specify packages with attributes",
but it reappeared in the 2012-11-05 commit for "Support for postgresql 9.0.x on
opensuse 11.4".

Clearly, the old code badly breaks the redhat family for postgresql-9.X, and
since it probably has no effect in opensuse, I just decided to remove it.