Run initdb for any redhat-derived platform version on first install
This should have been done in the 2012-11-02 commit to "[COOK-1844]
Make service name configurable." That change corrected the initdb
call with the version-specific service name, and it also removed
the comment about the code being not valid for 9.x. It should have
removed the "if node[postgresql][version].to_f < 9.0", too.

That is what I have done, to complete that fix.