Support Installing non-standard versions on CentOS 6
The old package selection, service name and data directory
logic was such that it would always install the basic
postgresql packages (8.4) on CentOS 6 even if the node's
postgresql version was set to be different to the CentOS 6
default version and the pgdg repository was enabled.

This commit adds extra logic for this new scenario without
(I believe) affecting any of the previous logic.