Fixes service from pgdg package on ubuntu 10.04
The pgdg repo's package names the service 'postgresql' instead of
'postgresql-VER'. The attributes file checks that the platform version
of ubuntu is 10.04 and appends the postgresql version. This causes the
service to fail in the debian_server recipe.

This commit fixes this by checking if the pgdg repo should be used.
Since this is false by default (set later in the file), it will only
evaluate true if the user sets it elsewhere like a role.