Update URL for PGDG Key
The maintainers have changed where the key is meant to be located In October/November 2013, and can be seen in their documentation here:
http://wiki.postgresql.org/index.php?title=Apt&diff=21465&oldid=21447

Yesterday, the key was removed from the repo, causing these calls to fail. I spoke to one of the maintainers via IRC, and he's re-added it to the original location in the meantime, but will eventually be removed.