On OpenSUSE 13.1, startup fails if when /var/lib/pgsql/data is non-empty
Changed the time when the configuration files are placed into the directory to be after the first start of the server.
Subsequently changed the notifies to :immediately, otherwise changes I'd made to these files would not have been taken into account. I believe this is better than making whoever's using the cookbook/recipe explicitly have to restart the service if they've made any changes to conf files.