Fix install order of contrib packages to fix pg_stat_statements issues.
pg_stat_statements requires the "shared_preload_libraries" setting to be
set. However, on the first install, this was writing to postgresql.conf
which forced an immediate start or restart of postgresql. But at that
point, the contrib packages weren't actually installed, so this would
cause postgresql to fail to startup (since, for example, the
"pg_stat_statements" extension trying to be preloaded didn't actually
exist yet). By shifting the installation of the contrib packages to
before the server, this allows the dependencies to be installed before
the server needs them.