Run restart command right away on the postgresql service.
This fixes errors when using docker, without it the error `psql: could not connect to server: no such file or directory` is displayed when this recipe is used with docker and chef-solo. This is the same issue that is affecting https://github.com/hw-cookbooks/postgresql/pull/206#issuecomment-63587051 however I feel firing 1 restart is better than firing 1 for each configuration change.