add attribute to use templates from other cookbook in server_conf.rb recipe
Added attributes which default to nil. When not nil, allows using
templates in recipes/server_conf.rb (e.g. for postgresql.conf,
pg_hba.conf) from another cookbook. This is simpler than requiring
the use of chef-rewind and allows for overriding the templates
when postgresql community cookbook is included in a LWRP (e.g.
with `run_context.include_recipe`), where `rewind` may not
behave as expected.