Improve how postgres password is set
Changed resource from bash to execute so that the sensitive property
can be used. Modified the not_if guard to actually check if it can
use the password to make a simply query. This check assumes that
pg_hba.conf forces a password for TCP connections. The password comes
from the PGPASSWORD environment variable.