Custom resource for installing extensions
The contrib recipe is very restrictive in that it only allows you to
install a single set of extensions into a single database for any
given node.

This new resource allows installation of any extension into any
database at any time. It also supports dropping extensions.

The name property takes the form "database/extension" to ensure that
resources remain unique when dealing with extensions across multiple
databases.

The call to psql now enables ON_ERROR_STOP to ensure that the
operation does not silently fail.

The contrib recipe has been kept but now utilises the resource. The
resource does not install postgresql-contrib so the recipe should
still be used for this.