Simply / do the right thing with the repos
When you do a client install just use the defaults for the version you want to install. If you want to customize how you setup the repo then you disable the auto repo creation and use the resource directly. Out of the box it works for most people and for those that need to self host the repo they can do so.

Signed-off-by: Tim Smith <tsmith@chef.io>