More work towards a resource driven cookbook (#458)
* Only test our new resources for now

* Add descriptions to the yum repos

Otherwise they're just Yum repository

* Do the right thing with the repos

When you do a client install just use the defaults for the version you want to install. If you want to customize how you setup the repo then you disable the auto repo creation and use the resource directly. Out of the box it works for most people and for those that need to self host the repo they can do so.

* Require Chef 12.14+ and remove SUSE support

We can add it back later, but there's not a postgresql.org zypper repo. They officially build via opensuse.org which we can use, but it's going to take some work.

* Update integration tests

* Change auto_repo_setup to setup_repo

* Speed up the repo

* Use same repo setup logic on the server side

* Update server spec

* Fix package installation on Debian systems

* Remove opensuse from Test Kitchen for now

* Install sudo for Travis

* Use the Fedora 26 box not latest

* Remove the config resource for now

* Turn of duplicate resource warning for the repo resource

Signed-off-by: Tim Smith <tsmith@chef.io>