Rip out service notification hack
This was a workaround to https://github.com/chef/chef/issues/6987 that
was proposed by @coderanger a while back. We use a resource action to
signal the parent resource so you can trigger notifications from the
custom resource. A delayed_action disconnects the notification system
because it waits to converge until the very end of the chef run and most
often is too late to properly send any notification messages.