Change initdb command to use the more explicit /bin/initdb version
Change teh default locale to C as everyone defintiely has that.

This allows us to pass in more options for locale and auth etc.
We currently only pass in locale, but this hould allow for further easy
customisation of the command. The command is also now the same for
systemd and non-systemd.

Fixes #536
Fixes #537