Provide ability to pass execution user from resources
resources -

    1. database
    2. user
    3. extensions

background -

It may possible that the machine which is going to run the recipe does
not have postgres user. Chef will automatically run the command eith root user
if we don't not specify the user.
We will lose the ability of peer authentication if we don't specify the user
so wecan provide the user name in params if you need to run sql commands using
specific user.