Fix idempotency when installing multiple client versions
This fixes an issue when trying to install multiple versions of postgresql on
the same system. This is useful on backup servers where you might be backing up
several different versions and it's best to use the correct version when doing
that.

On RHEL-based systems, this changes the name of the gpg key to *not* include the
PostgreSQL version. Each repo uses the same key so this shouldn't be a problem.

On Debian-based systems, create a different repo based on the version you're
using so that it also works on those systems.

Signed-off-by: Lance Albertson <lance@osuosl.org>