Cast `server_config` keys to strings in `postgresql_config` to avoid unnecessary converges (#782)
Reading the existing config on the node in `load_current_value` returns
string keys. If you then pass the same configuration but with symbol
keys to `server_config` in `postgresql_config`, `converge_if_changed`
will believe the resource changed and proceed with converging. The keys
are already cast to strings before being passed to the `template`
resource, so the template itself doesn't actually change but the
`postgresql_config` resource still believes it changed, so any
notifications will fire. For example, if you have the resource reload
postgresql on change, it'll reload each time the node is converged.

By casting the keys to strings when coercing the property, the resource
will no longer believe it changed when the only difference was the class
of the key, thus it wont fire notifications unnecessarily.

Signed-off-by: Chris Gunther <chris@room118solutions.com>