Use --platform ruby to ensure we install the source based pg gem (#791)
The upstream pg ruby gem has started including pre-complied pg gems which do not
work on EL8 due to the fact they were compiled on an Ubuntu platform that has a
newer glibc.

By using `--platform ruby` [1], this will force the installation using the
source which is what this cookbook expects.

In addition, make sure we test on EL8.

[1] https://github.com/ged/ruby-pg?tab=readme-ov-file#source-gem

Signed-off-by: Lance Albertson <lance@osuosl.org>