Ensure access config file is written before notifications are sent
If you use the postgresql_access resource and have it trigger a reload of the
postgresql_service resource, it will not do it in the correct order. Currently,
it will reload the server, and then the config file gets updated which means the
new config is never reloaded.

This ensures that it happens at the end of the resource run so it works in the
proper order.

Signed-off-by: Lance Albertson <lance@osuosl.org>